Feature add blocked

+++ src/users/users.controller.ts
@@ -13,6 +13,7 @@ import { CreateUserDto } from './dtos/create-user.dto';
 import { CreateAdminDto } from './dtos/create-admin.dto';
 import { UsersService } from './users.service';
 import { UpdateUserDto } from './dtos/update-user.dto';
+import { ChangeStatusUserDto } from './dtos/change-status-user';
 import { Serialize } from '../interceptors/serialize.interceptor';
 import { UserDto } from './dtos/user.dto';
 import { AuthService } from './auth.service';
@@ -41,7 +42,6 @@ export class UsersController {
       body.country,
       body.latitude,
       body.longitude,
-      body.profileimage,
     );
   }
 
@@ -75,12 +75,38 @@ export class UsersController {
   }
 
   @Delete('users/:id')
-  deleteUser(@Param('id') id: string) {
+  @UseGuards(AuthGuard)
+  deleteUser(@Param('id') id: string, @CurrentUser() user: User) {
+    console.log(user);
+    if (user.id != parseInt(id) && !user.isAdmin) {
+      throw new UnauthorizedException();
+    }
     return this.usersService.remove(parseInt(id));
   }
 
   @Patch('users/:id')
-  updateUser(@Param('id') id: string, @Body() body: UpdateUserDto) {
+  @UseGuards(AuthGuard)
+  updateUser(
+    @Param('id') id: string,
+    @Body() body: UpdateUserDto,
+    @CurrentUser() user: User,
+  ) {
+    if (user.id != parseInt(id) && !user.isAdmin) {
+      throw new UnauthorizedException();
+    }
+    return this.usersService.update(parseInt(id), body);
+  }
+
+  @Patch('users/:id/status')
+  @UseGuards(AuthGuard)
+  updateUserStatus(
+    @Param('id') id: string,
+    @Body() body: ChangeStatusUserDto,
+    @CurrentUser() user: User,
+  ) {
+    if (!user.isAdmin) {
+      throw new UnauthorizedException();
+    }
     return this.usersService.update(parseInt(id), body);
   }
 }
